[GREEN] add more validity checking on service

+++ src/main/java/id/ac/ui/cs/advprog/youkosoproduct/service/VoucherServiceImpl.java
@@ -25,6 +25,8 @@ public class VoucherServiceImpl implements VoucherService {
 
     @Override
     public void delete(Long voucherId) {
+        if (!voucherRepository.existsById(voucherId)) 
+            throw new IllegalArgumentException("There is no voucher with ID " + voucherId);
         voucherRepository.deleteById(voucherId);
     }
 
@@ -51,15 +53,24 @@ public class VoucherServiceImpl implements VoucherService {
             throw new IllegalArgumentException("There is no voucher with ID " + id);
 
         Voucher voucher = optionalVoucher.get();
+        
+        Voucher editedVoucher = new Voucher();
+        editedVoucher.setName(name);
+        editedVoucher.setDiscountPercentage(discountPercentage);
+        editedVoucher.setHasUsageLimit(hasUsageLimit);
+        editedVoucher.setUsageLimit(usageLimit);
+        editedVoucher.setMinimumOrder(minimumOrder);
+        editedVoucher.setMaximumDiscountAmount(maximumDiscountAmount);
+
+        if (!editedVoucher.isValid()) 
+            throw new IllegalArgumentException("Invalid voucher's attribute(s)");
+        
         voucher.setName(name);
         voucher.setDiscountPercentage(discountPercentage);
         voucher.setHasUsageLimit(hasUsageLimit);
         voucher.setUsageLimit(usageLimit);
         voucher.setMinimumOrder(minimumOrder);
         voucher.setMaximumDiscountAmount(maximumDiscountAmount);
-
-        if (!voucher.isValid())
-            throw new IllegalArgumentException("Wrong voucher's attribute(s)");
 
         voucherRepository.save(voucher);
     }
